Add files via upload
the register html file is used to feed data into a db named users.
the login html file validates the entered info and reacts accordingly. 
This is NOT the final version, im having some compatibility issues on my computer at the moment. Will get back to it in a while.

NOTE - for these files to be called, the front_page.html file has the be modified but i havent committed that as of now.